backend/scrapers/award_scraper/requirements.txt
fastapi
uvicorn
playwright
google-cloud-pubsub
pydantic
